pyproject.toml
[tool.poetry]
name = "espire-music"
version = "0.1.0"
description = ""
authors = ["kurt4ins <zhenya_arz@bk.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
pillow = "^10.4.0"
djangorestframework = {extras = ["json"], version = "^3.15.2"}
django-cors-headers = "^4.4.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
